pyproject.toml
[project]
name = "rag-doc-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-search-documents>=11.5.2",
    "fastapi[standard]>=0.115.11",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.45",
    "langchain-text-splitters>=0.3.6",
    "pymongo>=4.11.3",
    "azure-keyvault-secrets>=4.9.0",
    "azure-core>=1.32.0",
    "azure-storage-blob>=12.25.1",
    "openai>=1.69.0",
    "azure-keyvault>=4.2.0",
    "libmagic>=1.0",
    "pymupdf>=1.25.5",
    "python-magic>=0.4.27",
    "uvicorn>=0.34.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "mkdocs>=1.6.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.10",
    "rust-just>=1.40.0",
    "python-dotenv>=1.0.1",
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
